- Implementation for update and delete fixed such that if the operation is performed on Candidate that doesn't exist then 0 is returned back to the client  - Problem is the queryByObject throws exceptions if any other number than 1 is returned by the query (wow, so fragile)  - Replaced the query and then made a new exception for a duplicate primary query since regular queries can return 0 or more rows, and if it does for a getbyid then that means that multiple entities have the same id, which means there is a duplicate primary key in the database, in which case something terrible has happened.

+++ webapi/src/main/java/com/webapi/webapi/api/controllers/CandidateController.java
@@ -1,5 +1,6 @@
 package com.webapi.webapi.api.controllers;
 
+import com.webapi.webapi.databasedrivers.DuplicatePrimaryKeyException;
 import com.webapi.webapi.model.candidate.Candidate;
 import com.webapi.webapi.model.candidate.NonExistentCandidateException;
 import com.webapi.webapi.services.CandidateService;
@@ -26,7 +27,7 @@ public class CandidateController {
     }
 
     @GetMapping("/{id}")
-    public Candidate get(@PathVariable Long id) throws NonExistentCandidateException {
+    public Candidate get(@PathVariable Long id) throws NonExistentCandidateException, DuplicatePrimaryKeyException {
         return candidateService.get(id);
     }
 
@@ -46,8 +47,11 @@ public class CandidateController {
     }
 
     @DeleteMapping("/{id}")
-    public int delete(@PathVariable Long id) throws NonExistentCandidateException {
+    public int delete(@PathVariable Long id) throws NonExistentCandidateException, DuplicatePrimaryKeyException {
         Candidate candidateToDelete = candidateService.get(id);
+        if (candidateToDelete == null) {
+            return 0;
+        }
         return candidateService.delete(candidateToDelete);
     }
